==> 1)_data_capture.ipynb <==
# My repo
!git clone https://github.com/Tixonmavrin/covid-19-solution
# Data world
!git clone https://github.com/CSSEGISandData/COVID-19.git
# Data world new
!git clone https://github.com/CSSEGISandData/COVID-19 -b web-data COVID-19-new
# Data Russia
!git clone https://github.com/grwlf/COVID-19_plus_Russia.git
# More data
!wget https://coronadatascraper.com/timeseries.csv

import pandas as pd
import requests
import numpy as np

# World confirmed
data = pd.read_csv("COVID-19/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_global.csv")
data['Country/Region'] = data['Country/Region'].apply(lambda x: str(x))
data['Province/State'] = data['Province/State'].apply(lambda x: str(x))
data = data.groupby(["Country/Region", "Province/State"]).sum().reset_index()
del data["Lat"]
del data["Long"]
data.columns = list(data.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in data.columns[2:]]

# World deaths
data_deaths = pd.read_csv("COVID-19/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_global.csv")
data_deaths['Country/Region'] = data_deaths['Country/Region'].apply(lambda x: str(x))
data_deaths['Province/State'] = data_deaths['Province/State'].apply(lambda x: str(x))
data_deaths = data_deaths.groupby(["Country/Region", "Province/State"]).sum().reset_index()
del data_deaths["Lat"]
del data_deaths["Long"]
data_deaths.columns = list(data_deaths.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in data_deaths.columns[2:]]

# US confirmed
data_US = pd.read_csv("COVID-19/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_US.csv")
for col in ["UID", "iso2","iso3","code3","FIPS","Admin2", "Combined_Key"]:
  del data_US[col]
data_US = data_US.rename(columns={"Long_":"Long", "Country_Region":"Country/Region","Province_State":"Province/State"})
data_US['Country/Region'] = data_US['Country/Region'].apply(lambda x: str(x))
data_US['Province/State'] = data_US['Province/State'].apply(lambda x: str(x))
data_US = data_US.groupby(["Country/Region", "Province/State"]).sum().reset_index()
del data_US["Lat"]
del data_US["Long"]
data_US.columns = list(data_US.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in data_US.columns[2:]]

# US deaths
data_deaths_US = pd.read_csv("COVID-19/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_US.csv")
for col in ["UID", "iso2","iso3","code3","FIPS","Admin2", "Combined_Key", "Population"]:
  del data_deaths_US[col]
data_deaths_US = data_deaths_US.rename(columns={"Long_":"Long", "Country_Region":"Country/Region","Province_State":"Province/State"})
data_deaths_US['Country/Region'] = data_deaths_US['Country/Region'].apply(lambda x: str(x))
data_deaths_US['Province/State'] = data_deaths_US['Province/State'].apply(lambda x: str(x))
data_deaths_US = data_deaths_US.groupby(["Country/Region", "Province/State"]).sum().reset_index()
del data_deaths_US["Lat"]
del data_deaths_US["Long"]
data_deaths_US.columns = list(data_deaths_US.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in data_deaths_US.columns[2:]]

data = data.append(data_US)
data_deaths = data_deaths.append(data_deaths_US)

# Today data countries
data_today = pd.read_csv("COVID-19-new/data/cases_country.csv")

data_today_deaths = data_today[["Country_Region", "Lat", "Long_", "Deaths"]]
data_today_deaths["Province/State"] = 'nan'
data_today_deaths = data_today_deaths.rename(columns={"Long_":"Long", "Deaths":pd.to_datetime('today').strftime("%m/%d/%y"), "Country_Region":"Country/Region"})
data_today_deaths['Country/Region'] = data_today_deaths['Country/Region'].apply(lambda x: str(x))
data_today_deaths['Province/State'] = data_today_deaths['Province/State'].apply(lambda x: str(x))
data_today_deaths = data_today_deaths.groupby(["Country/Region", "Province/State"]).sum().reset_index()
del data_today_deaths["Lat"]
del data_today_deaths["Long"]
data_today_deaths.columns = list(data_today_deaths.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in data_today_deaths.columns[2:]]

data_today_confirmed = data_today[["Country_Region", "Lat", "Long_", "Confirmed"]]
data_today_confirmed["Province/State"] = 'nan'
data_today_confirmed = data_today_confirmed.rename(columns={"Long_":"Long", "Confirmed":pd.to_datetime('today').strftime("%m/%d/%y"), "Country_Region":"Country/Region"})
data_today_confirmed['Country/Region'] = data_today_confirmed['Country/Region'].apply(lambda x: str(x))
data_today_confirmed['Province/State'] = data_today_confirmed['Province/State'].apply(lambda x: str(x))
data_today_confirmed = data_today_confirmed.groupby(["Country/Region", "Province/State"]).sum().reset_index()
del data_today_confirmed["Lat"]
del data_today_confirmed["Long"]
data_today_confirmed.columns = list(data_today_confirmed.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in data_today_confirmed.columns[2:]]

# Today data states
states_today = pd.read_csv("COVID-19-new/data/cases_state.csv")

states_today_deaths = states_today[["Country_Region", "Province_State", "Lat", "Long_", "Deaths"]]
states_today_deaths = states_today_deaths.rename(columns={"Long_":"Long", "Deaths":pd.to_datetime('today').strftime("%m/%d/%y"), "Country_Region":"Country/Region", "Province_State":"Province/State"})
states_today_deaths['Country/Region'] = states_today_deaths['Country/Region'].apply(lambda x: str(x))
states_today_deaths['Province/State'] = states_today_deaths['Province/State'].apply(lambda x: str(x))
states_today_deaths = states_today_deaths.groupby(["Country/Region", "Province/State"]).sum().reset_index()
del states_today_deaths["Lat"]
del states_today_deaths["Long"]
states_today_deaths.columns = list(states_today_deaths.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in states_today_deaths.columns[2:]]

states_today_confirmed = states_today[["Country_Region", "Province_State", "Lat", "Long_", "Confirmed"]]
states_today_confirmed = states_today_confirmed.rename(columns={"Long_":"Long", "Confirmed":pd.to_datetime('today').strftime("%m/%d/%y"), "Country_Region":"Country/Region", "Province_State":"Province/State"})
states_today_confirmed['Country/Region'] = states_today_confirmed['Country/Region'].apply(lambda x: str(x))
states_today_confirmed['Province/State'] = states_today_confirmed['Province/State'].apply(lambda x: str(x))
states_today_confirmed = states_today_confirmed.groupby(["Country/Region", "Province/State"]).sum().reset_index()
del states_today_confirmed["Lat"]
del states_today_confirmed["Long"]
states_today_confirmed.columns = list(states_today_confirmed.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in states_today_confirmed.columns[2:]]

data_today_confirmed = data_today_confirmed.append(states_today_confirmed)
data_today_deaths = data_today_deaths.append(states_today_deaths)

data = data.merge(data_today_confirmed, how='left', on=['Country/Region','Province/State'])
data_deaths = data_deaths.merge(data_today_deaths, how='left', on=['Country/Region','Province/State'])
data.to_csv('covid-19-solution/data/data_start/_data_confirmed.csv', index=False)
data_deaths.to_csv('covid-19-solution/data/data_start/_data_deaths.csv', index=False)
data.sample(3)

# Russia confirmed
data_rus = pd.read_csv("COVID-19_plus_Russia/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_confirmed_RU.csv")
for col in ["UID", "iso2","iso3","code3","FIPS","Admin2", "Combined_Key"]:
  del data_rus[col]
data_rus = data_rus.rename(columns={"Long_":"Long", "Country_Region":"Country/Region","Province_State":"Province/State"})
data_rus['Country/Region'] = data_rus['Country/Region'].apply(lambda x: str(x))
data_rus['Province/State'] = data_rus['Province/State'].apply(lambda x: str(x))
data_rus = data_rus.groupby(["Country/Region", "Province/State"]).sum().reset_index()
data_rus['Province/State'] = data_rus['Province/State'].apply(lambda x: x if (x != 'Altay republic') else 'Republic of Altay')
del data_rus["Lat"]
del data_rus["Long"]
data_rus.columns = list(data_rus.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in data_rus.columns[2:]]

# Russia deaths
data_deaths_rus = pd.read_csv("COVID-19_plus_Russia/csse_covid_19_data/csse_covid_19_time_series/time_series_covid19_deaths_RU.csv")
for col in ["UID", "iso2","iso3","code3","FIPS","Admin2", "Combined_Key"]:
  del data_deaths_rus[col]
data_deaths_rus = data_deaths_rus.rename(columns={"Long_":"Long", "Country_Region":"Country/Region","Province_State":"Province/State"})
data_deaths_rus['Country/Region'] = data_deaths_rus['Country/Region'].apply(lambda x: str(x))
data_deaths_rus['Province/State'] = data_deaths_rus['Province/State'].apply(lambda x: str(x))
data_deaths_rus = data_deaths_rus.groupby(["Country/Region", "Province/State"]).sum().reset_index()
data_deaths_rus['Province/State'] = data_deaths_rus['Province/State'].apply(lambda x: x if (x != 'Altay republic') else 'Republic of Altay')
del data_deaths_rus["Lat"]
del data_deaths_rus["Long"]
data_deaths_rus.columns = list(data_deaths_rus.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in data_deaths_rus.columns[2:]]

# Data today (from coronavirus-monitor.ru)
class window():
  pass
true = True
false = False
null = np.nan
response = requests.get('https://coronavirus-monitor.ru/jquery-lite-9.js')
result = response.text.strip()
exec(result)
date = window.dataFromServer['russianSubjects']['data']['subjects'][0]['statistics'][-1]['date']

countries_dict = {
' Kabardino-Balkarian Republic': 'Republic of Kabardino-Balkaria',' Kurgan region': 'Kurgan oblast',
'Adygeya Republic': 'Republic of Adygeia','Altai Krai': 'Altayskiy kray',
'Amur oblast': 'Amursk oblast','Arkhangelsk region': 'Arkhangelsk oblast',
'Astrakhan oblast': 'Astrahan oblast','Bashkortostan': 'Republic of Bashkortostan',
'Chechen Republic ': 'Chechen republic','Chelyabinsk region': 'Cheliabinsk oblast',
'Chukotka Autonomous Okrug': 'Chukotskiy autonomous oblast','Chuvash Republic': 'Republic of Chuvashia',
'Irkutsk oblast': 'Irkutsk oblast','Ivanovo region': 'Ivanovo oblast',
'Jewish autonomous oblast': 'Jewish Autonomous oblast','Kaluga region': 'Kaluga oblast',
'Kamchatka Krai': 'Kamchatskiy kray','Karachay-Cherkess Republic': 'Republic of Karachaevo-Cherkessia',
'Khabarovsk territory': 'Habarovskiy kray','Khanty-Mansi Autonomous Okrug': 'Hanty-Mansiyskiy AO',
'Kirov region': 'Kirov oblast','Kostroma oblast': 'Kostroma oblast',
'Krasnodar region': 'Krasnodarskiy kray','Krasnoyarsk region': 'Krasnoyarskiy kray',
'Kursk oblast': 'Kursk oblast','Leningrad region': 'Leningradskaya oblast',
'Magadan oblast': 'Magadan oblast','Moscow': 'Moscow',
'Moscow oblast': 'Moscow oblast','Murmansk region': 'Murmansk oblast',
'Nenets Autonomous Okrug': 'Nenetskiy autonomous oblast','Nizhny Novgorod region': 'Nizhegorodskaya oblast',
'Novgorod region': 'Novgorod oblast','Novosibirsk region': 'Novosibirsk oblast',
'Omsk oblast': 'Omsk oblast','Orel region': 'Orel oblast',
'Orenburg region': 'Orenburg oblast','Penza region': 'Pensa oblast',
'Perm region': 'Perm oblast','Pskov Oblast': 'Pskov oblast',
'Republic Of Khakassia': 'Republic of Hakassia','Republic of Buryatia': 'Republic of Buriatia',
'Republic of Crimea': 'Republic of Crimea','Republic of Dagestan': 'Republic of Dagestan',
'Republic of Kalmykia': 'Republic of Kalmykia','Republic of Karelia': 'Republic of Karelia',
'Republic of Mari El': 'Republic of Mariy El','Republic of Mordovia': 'Republic of Mordovia',
'Rostov oblast': 'Rostov oblast','Ryazan region': 'Ryazan oblast',
'Saint Petersburg': 'Saint Petersburg','Sakhalin oblast': 'Sakhalin oblast',
'Samara region': 'Samara oblast','Saratov region': 'Saratov oblast',
'Sevastopol': 'Sevastopol','Smolensk oblast': 'Smolensk oblast',
'Sverdlovsk region': 'Sverdlov oblast','Tambov region': 'Tambov oblast',
'The Republic of Ingushetia': 'Ingushetia republic','The Republic of North Ossetia – Alania ': 'Republic of North Osetia - Alania',
'Tomsk region': 'Tomsk oblast','Transbaikal territory': 'Zabaykalskiy kray',#
'Tula region': 'Tula oblast','Tuva Republic': 'Republic of Tyva',
'Tver region': 'Tver oblast','Tyumen region': 'Tumen oblast',
'Udmurt Republic': 'Republic of Udmurtia','Ulyanovsk region': 'Ulianovsk oblast',
'Vladimir region': 'Vladimir oblast','Volgograd oblast': 'Volgograd oblast',
'Vologda oblast': 'Vologda oblast','Voronezh region': 'Voronezh oblast',
'Yakutia': 'Saha republic','Yamalo-Nenets Autonomous district': 'Yamalo-Nenetskiy AO',
'Yaroslavl oblast': 'Yaroslavl oblast','Белгородская область': 'Belgorod oblast',
'Брянская область': 'Briansk oblast','Калининградская область': 'Kaliningrad oblast',
'Кемеровская область':  'Kemerovo oblast','Липецкая область': 'Lipetsk oblast',
'Приморский край': 'Primorskiy kray','Республика Коми': 'Komi republic',
'Ставропольский край': 'Stavropolskiy kray','Татарстан': 'Republic of Tatarstan',
'Республика Алтай': 'Republic of Altay',
}

dict_russia_conf = {}
dict_russia_dea = {}
for v in window.dataFromServer['russianSubjects']['data']['subjects']:
  dict_russia_conf[v['en']] = v['statistics'][-1]['confirmed']
  dict_russia_dea[v['en']] = v['statistics'][-1]['deaths']

# Confirmed
dict_russia_conf_main = []
for n in dict_russia_conf:
  dict_russia_conf_main.append([countries_dict[n], dict_russia_conf[n]])
dict_russia_conf_main = pd.DataFrame(np.array(dict_russia_conf_main))
dict_russia_conf_main.columns = ['Province/State',pd.to_datetime('today').strftime("%m/%d/%y")]
dict_russia_conf_main['Country/Region'] = 'Russia'
dict_russia_conf_main['Country/Region'] = dict_russia_conf_main['Country/Region'].apply(lambda x: str(x))
dict_russia_conf_main['Province/State'] = dict_russia_conf_main['Province/State'].apply(lambda x: str(x))
dict_russia_conf_main = dict_russia_conf_main.groupby(["Country/Region", "Province/State"]).sum().reset_index()
dict_russia_conf_main.columns = list(dict_russia_conf_main.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in dict_russia_conf_main.columns[2:]]

# Deaths
dict_russia_dea_main = []
for n in dict_russia_dea:
  dict_russia_dea_main.append([countries_dict[n], dict_russia_dea[n]])
dict_russia_dea_main = pd.DataFrame(np.array(dict_russia_dea_main))
dict_russia_dea_main.columns = ['Province/State',pd.to_datetime('today').strftime("%m/%d/%y")]
dict_russia_dea_main['Country/Region'] = 'Russia'
dict_russia_dea_main['Country/Region'] = dict_russia_dea_main['Country/Region'].apply(lambda x: str(x))
dict_russia_dea_main['Province/State'] = dict_russia_dea_main['Province/State'].apply(lambda x: str(x))
dict_russia_dea_main = dict_russia_dea_main.groupby(["Country/Region", "Province/State"]).sum().reset_index()
dict_russia_dea_main.columns = list(dict_russia_dea_main.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in dict_russia_dea_main.columns[2:]]

dict_russia_conf_main[dict_russia_conf_main.columns[-1]] = dict_russia_conf_main[dict_russia_conf_main.columns[-1]].astype('int64')
dict_russia_dea_main[dict_russia_dea_main.columns[-1]] = dict_russia_dea_main[dict_russia_dea_main.columns[-1]].astype('int64')

data_rus = data_rus.merge(dict_russia_conf_main, how='left', on=['Country/Region','Province/State'])
data_deaths_rus = data_deaths_rus.merge(dict_russia_dea_main, how='left', on=['Country/Region','Province/State'])
data_rus.to_csv('covid-19-solution/data/data_start/_data_confirmed_rus.csv', index=False)
data_deaths_rus.to_csv('covid-19-solution/data/data_start/_data_deaths_rus.csv', index=False)
data_rus.sample(3)

pd.read_csv('timeseries.csv').to_csv('covid-19-solution/data/other/timeseries.csv')
pd.read_csv('https://coronadatascraper.com/timeseries-jhu.csv').to_csv('covid-19-solution/data/other/timeseries-jhu.csv')

data_all = pd.read_csv('https://coronadatascraper.com/timeseries-jhu.csv')
data_all.drop(['level', 'city', 'county', 'state', 'lat', 'long', 'population', 'url', 'aggregate', 'tz'], axis=1, inplace=True)
data_all = data_all.rename(columns={'country':'Country/Region', 'name':'Province/State'})
data_all.columns = list(data_all.columns[:2]) + [pd.to_datetime(i).strftime("%m/%d/%y") for i in data_all.columns[2:]]

data_all.shape

data_all2 = []
data_all_cols = data_all.columns
for v in data_all.values:
  if len(np.unique(v[2:])) > 40:
    data_all2.append(v)
data_all2 = pd.DataFrame(data_all2)
data_all2.columns = data_all_cols

data_all2.shape

data2 = data.append(data_all2)
data2 = data2.fillna(0)
data2.to_csv('covid-19-solution/data/data_start/_data_confirmed_all.csv')
