==> 3)_make_features.ipynb <==
# My repo with data
!git clone https://github.com/Tixonmavrin/covid-19-solution
!unzip covid-19-solution/data/other/timeseries.csv.zip

import pandas as pd
import numpy as np

#data = pd.read_csv('covid-19-solution/data/data_clean/_nonoice_data_confirmed.csv')
#data_deaths = pd.read_csv('covid-19-solution/data/data_clean/_nonoice_data_deaths.csv')
#data_rus = pd.read_csv('covid-19-solution/data/data_clean/_nonoice_data_confirmed_rus.csv')
#data_deaths_rus = pd.read_csv('covid-19-solution/data/data_clean/_nonoice_data_deaths_rus.csv')

data = pd.read_csv('covid-19-solution/data/data_start/_data_confirmed.csv')
data_deaths = pd.read_csv('covid-19-solution/data/data_start/_data_deaths.csv')
data_rus = pd.read_csv('covid-19-solution/data/data_start/_data_confirmed_rus.csv')
data_deaths_rus = pd.read_csv('covid-19-solution/data/data_start/_data_deaths_rus.csv')

data = pd.read_csv('covid-19-solution/data/data_start/_data_confirmed_all.csv')
del data['Unnamed: 0']
data['Province/State'] = data['Province/State'].fillna('nan')

data

data = data[(data['Province/State'] != 'Diamond Princess') & (data['Province/State'] != 'Grand Princess')]
data_deaths = data_deaths[(data_deaths['Province/State'] != 'Diamond Princess') & (data_deaths['Province/State'] != 'Grand Princess')]

# Change format

#data = data.append(data_rus)
#data_deaths = data_deaths.append(data_deaths_rus)
data = data.groupby(["Country/Region","Province/State"]).sum()
data = data.stack().reset_index()
data.columns = ["Country/Region","Province/State","Date","Confirmed"]

data_deaths = data_deaths.groupby(["Country/Region","Province/State"]).sum()
data_deaths = data_deaths.stack().reset_index()
data_deaths.columns = ["Country/Region","Province/State","Date","Deaths"]

data_rus = data_rus.groupby(["Country/Region","Province/State"]).sum()
data_rus = data_rus.stack().reset_index()
data_rus.columns = ["Country/Region","Province/State","Date","Confirmed"]

data_deaths_rus = data_deaths_rus.groupby(["Country/Region","Province/State"]).sum()
data_deaths_rus = data_deaths_rus.stack().reset_index()
data_deaths_rus.columns = ["Country/Region","Province/State","Date","Deaths"]

data = data.merge(data_deaths, how='left', on=['Country/Region','Province/State','Date'])
data_rus = data_rus.merge(data_deaths_rus, how='left', on=['Country/Region','Province/State','Date'])

data.Date = pd.to_datetime(data.Date)
data.Confirmed = data.Confirmed.fillna(0).astype('int64')
data.Deaths = data.Deaths.fillna(0).astype('int64')

data_rus.Date = pd.to_datetime(data_rus.Date)
data_rus.Confirmed = data_rus.Confirmed.astype('int64')
data_rus.Deaths = data_rus.Deaths.astype('int64')

# You can get new data

main_timeseries = pd.read_csv('timeseries.csv')
main_timeseries = main_timeseries[['country','name','date','cases', 'deaths','level']]
main_timeseries = main_timeseries[main_timeseries['level'] == 'state']
del main_timeseries['level']
main_timeseries.columns = ['Country/Region','Province/State', 'Date', 'Confirmed', 'Deaths']
main_timeseries['Confirmed'] = main_timeseries['Confirmed'].fillna(0)
main_timeseries['Deaths'] = main_timeseries['Deaths'].fillna(0)
#data = data.append(main_timeseries)

# Countries and regions info

russia_regions = pd.read_csv("covid-19-solution/data/features/russia_regions.csv")
russia_regions['urban_pop_rate']=russia_regions['population_urban']/russia_regions['population']
russia_regions = russia_regions[['csse_province_state', 'population','population_urban','population_rural','urban_pop_rate', 'geoname_name']]
russia_regions.loc[russia_regions['geoname_name'] == 'Nenets', 'csse_province_state'] = 'Nenetskiy autonomous oblast'
russia_regions.loc[russia_regions['geoname_name'] == 'Chukotka', 'csse_province_state'] = 'Chukotskiy autonomous oblast'
russia_regions.loc[russia_regions['geoname_name'] == 'Altayskiy kray', 'csse_province_state'] = 'Altay republic'
del russia_regions['geoname_name']
russia_regions = russia_regions.rename({'csse_province_state': 'Province/State'})

qd=pd.read_csv("covid-19-solution/data/mobility/quarantine.csv")
qd=qd.append({'Start date': '2020-03-19','Country':'USA','Level':'National'}, ignore_index=True)
qd=qd.append({'Start date': '2020-03-16','Country':'CZE','Level':'National'}, ignore_index=True)
qd = qd[['Country', 'Start date']]
qd['Start date'] = pd.to_datetime(qd['Start date'])
qd=qd.groupby(['Country']).min().reset_index()
qd.columns = ["Country/Region", "Start_date"]

Health_systems = pd.read_csv("covid-19-solution/data/features/2.12_Health_systems.csv").groupby("Country_Region").mean()
Health_systems = Health_systems.fillna(Health_systems.mean())
Health_systems = Health_systems[["Health_exp_pct_GDP_2016","Health_exp_per_capita_USD_2016","Physicians_per_1000_2009-18","Nurse_midwife_per_1000_2009-18"]]
Health_systems = Health_systems.reset_index()
Health_systems.columns = ["Country/Region","Health_GDP","Health_USD","Physicians","Nurse"]

WorldPop = pd.read_csv("covid-19-solution/data/features/WorldPopulationByAge2020.csv")
WorldPop = WorldPop.fillna(WorldPop.mean())
WorldPop2 = (WorldPop[WorldPop["AgeGrp"]=="60+"]["PopTotal"].to_numpy() + WorldPop[WorldPop["AgeGrp"]=="40-59"]["PopTotal"].to_numpy())/(WorldPop[WorldPop["AgeGrp"]=="0-19"]["PopTotal"].to_numpy()+WorldPop[WorldPop["AgeGrp"]=="20-39"]["PopTotal"].to_numpy())
WorldPop = WorldPop.groupby("Location").mean()
WorldPop["Age_old/new"] = WorldPop2
WorldPop = WorldPop[["Age_old/new"]].reset_index()
WorldPop.columns = ['Country/Region', 'Age_old/new']

smoke = pd.read_csv("covid-19-solution/data/features/share-of-adults-who-smoke.csv")
smoke = smoke[smoke.Year > 2015].groupby("Entity").mean()
smoke = smoke[["Smoking prevalence, total (ages 15+) (% of adults)"]]
smoke.columns = ["Smoking"]
smoke = smoke.reset_index()
smoke.columns = ['Country/Region', 'Smoking']

countries2 = pd.read_csv("covid-19-solution/data/features/covid19countryinfo.csv")
countries2 = countries2.fillna(0).groupby("country").sum()
countries2 = countries2[["tests","testpop","gatheringlimit","hospibed","healthperpop"]]
countries2 = countries2.reset_index()
countries2.columns = ['Country/Region', 'tests', 'testpop', 'gatheringlimit', 'hospibed','healthperpop']

countries = pd.read_csv("covid-19-solution/data/features/countries.csv")
countries.drop(['iso_alpha2','iso_numeric','official_name','name'], axis='columns',inplace=True)
del countries['iso_alpha3']
countries.columns = ['Country/Region', 'density_n', 'fertility_rate', 'land_area',
       'median_age', 'migrants', 'population_coun', 'urban_pop_rate_c', 'world_share']

russia_regions = russia_regions.rename(columns={'csse_province_state':'Province/State'})
data2 = data.merge(Health_systems, how='left', on='Country/Region')
data2 = data2.merge(WorldPop, how='left', on='Country/Region')
data2 = data2.merge(smoke, how='left', on='Country/Region')
data2 = data2.merge(countries2, how='left', on='Country/Region')
data2 = data2.merge(countries, how='left', on='Country/Region')
data2 = data2.merge(qd, how='left', on='Country/Region')
data2 = data2.merge(russia_regions, how='left', on='Province/State')

data_rus2 = data_rus.merge(Health_systems, how='left', on='Country/Region')
data_rus2 = data_rus2.merge(WorldPop, how='left', on='Country/Region')
data_rus2 = data_rus2.merge(smoke, how='left', on='Country/Region')
data_rus2 = data_rus2.merge(countries2, how='left', on='Country/Region')
data_rus2 = data_rus2.merge(countries, how='left', on='Country/Region')
data_rus2 = data_rus2.merge(qd, how='left', on='Country/Region')
data_rus2 = data_rus2.merge(russia_regions, how='left', on='Province/State')

region_metadata = pd.read_csv('covid-19-solution/data/features/region_metadata.csv')
del region_metadata['density']
del region_metadata['population']
region_metadata = region_metadata.rename(columns={'Country_Region':'Country/Region', 'Province_State':'Province/State'})
region_metadata['Province/State'] = region_metadata['Province/State'].fillna('nan')
data2 = data2.merge(region_metadata, how='left', on=['Country/Region', 'Province/State'])

del region_metadata['Province/State']
data_rus2 = data_rus2.merge(region_metadata, how='left', on='Country/Region')

countries = pd.read_csv('covid-19-solution/data/features/countries.csv')
m_apple = pd.read_csv('covid-19-solution/data/mobility/mobility-apple.csv')
m_apple = m_apple[['date', 'apl-driving', 'apl-transit', 'apl-walking',
       'name']]
m_apple.columns = ['Date', 'apl-driving', 'apl-transit', 'apl-walking',
       'Country/Region']
m_apple['Date'] = pd.to_datetime(m_apple['Date'])
m_apple.sample()

countries = countries.rename(columns={'ccse_name':'country'})

m_google = pd.read_csv("covid-19-solution/data/mobility/mobility-google.csv")

'''countries = pd.read_csv("covid-19-solution/data/features/countries.csv")
countries.drop(['iso_alpha2','iso_numeric','official_name','name'], axis='columns',inplace=True)
del countries['iso_alpha3']
countries.columns = ['Country/Region', 'density_n', 'fertility_rate', 'land_area',
       'median_age', 'migrants', 'population_coun', 'urban_pop_rate_c', 'world_share']
       '''
m_google = m_google.merge(countries[['country','name']], how='left', on='country')
del m_google['country']
m_google.columns = ['Date','grocery-and-pharmacy', 'parks', 'residential', 'retail-and-recreation',
       'transit-stations', 'workplaces', 'Country/Region']
m_google['Date'] = pd.to_datetime(m_google['Date'])
m_google.sample()

russia_regions = pd.read_csv('covid-19-solution/data/features/russia_regions.csv')[['iso_code', 'name', 'type', 'name_with_type', 'csse_province_state',
       'federal_district', 'timezone', 'geoname_code', 'geoname_id',
       'geoname_name', 'population', 'population_urban', 'population_rural']]
russia_regions.columns = ['country', 'name', 'type', 'name_with_type', 'Province/State',
       'federal_district', 'timezone', 'geoname_code', 'geoname_id',
       'geoname_name', 'population_r', 'population_urban', 'population_rural']

m_yandex = pd.read_csv("covid-19-solution/data/mobility/mobility-yandex.csv")
m_yandex = m_yandex.merge(russia_regions, how='left', on=['country'])
del m_yandex['country']
m_yandex['Country/Region'] = 'Russia'
m_yandex.columns = ['Date', 'isolation', 'name', 'type', 'name_with_type', 'Province/State',
       'federal_district', 'timezone', 'geoname_code', 'geoname_id',
       'geoname_name', 'population', 'population_urban', 'population_rural',
       'Country/Region']
m_yandex['Date'] = pd.to_datetime(m_yandex['Date'])
m_yandex = m_yandex[['Date', 'isolation', 'Province/State',
       'federal_district', 'geoname_name',
       'Country/Region']]
m_yandex.sample()

data3 = data2.merge(m_apple, how='left', on=['Country/Region', 'Date'])
data3 = data3.merge(m_google, how='left', on=['Country/Region', 'Date'])
data3 = data3.merge(m_yandex, how='left', on=['Country/Region', 'Province/State', 'Date'])

data_rus3 = data_rus2.merge(m_apple, how='left', on=['Country/Region', 'Date'])
data_rus3 = data_rus3.merge(m_google, how='left', on=['Country/Region', 'Date'])
data_rus3 = data_rus3.merge(m_yandex, how='left', on=['Country/Region', 'Province/State', 'Date'])

data4 = data3.append(data_rus3)

data4['Health_GDP'] = data4['Health_GDP'].fillna(data4['Health_GDP'].mean())
data4['Health_USD'] = data4['Health_USD'].fillna(data4['Health_USD'].mean())
data4['Physicians'] = data4['Physicians'].fillna(data4['Physicians'].mean())
data4['Nurse'] = data4['Nurse'].fillna(data4['Nurse'].mean())
data4['Age_old/new'] = data4['Age_old/new'].fillna(data4['Age_old/new'].mean())

data4['Smoking'] = data4['Smoking'].fillna(data4['Smoking'].mean())
data4['tests'] = data4['tests'].fillna(data4['tests'].mean())
data4['testpop'] = data4['testpop'].fillna(data4['testpop'].mean())
data4['gatheringlimit'] = data4['gatheringlimit'].fillna(data4['gatheringlimit'].mean())
data4['healthperpop'] = data4['healthperpop'].fillna(data4['healthperpop'].mean())
data4['density_n'] = data4['density_n'].fillna(data4['density_n'].mean())
data4['fertility_rate'] = data4['fertility_rate'].fillna(data4['fertility_rate'].mean())

data4['land_area'] = data4['land_area'].fillna(data4['land_area'].mean())
data4['median_age'] = data4['median_age'].fillna(data4['median_age'].mean())
data4['migrants'] = data4['migrants'].fillna(data4['migrants'].mean())
data4['population_coun'] = data4['population_coun'].fillna(data4['population_coun'].mean())
data4['urban_pop_rate'] = data4['urban_pop_rate'].fillna(data4['urban_pop_rate'].mean())
data4['world_share'] = data4['world_share'].fillna(data4['world_share'].mean())
data4['hospibed'] = data4['hospibed'].fillna(data4['hospibed'].mean())

data4.loc[data4.hospibed == 0,"hospibed"] = data4.hospibed.mean()
data4.loc[data4.tests == 0,"tests"] = data4.tests.mean()
data4.loc[data4.testpop == 0,"testpop"] = data4.testpop.mean()
data4.loc[data4.gatheringlimit == 0,"gatheringlimit"] = data4.gatheringlimit.mean()
data4.loc[data4.hospibed == 0,"hospibed"] = data4.hospibed.mean()

data4['population'] = data4['population'].fillna(data4['population'].mean())
data4['population_urban'] = data4['population_urban'].fillna(data4['population_urban'].mean())
data4['population_rural'] = data4['population_rural'].fillna(data4['population_rural'].mean())
data4['urban_pop_rate'] = data4['urban_pop_rate'].fillna(data4['urban_pop_rate'].mean())

data4['apl-driving'] = data4['apl-driving'].fillna(data4['apl-driving'].mean())
data4['apl-transit'] = data4['apl-transit'].fillna(data4['apl-transit'].mean())
data4['apl-walking'] = data4['apl-walking'].fillna(data4['apl-walking'].mean())
data4['isolation'] = data4['isolation'].fillna(data4['isolation'].mean())
data4['grocery-and-pharmacy'] = data4['grocery-and-pharmacy'].fillna(data4['grocery-and-pharmacy'].mean())
data4['parks'] = data4['parks'].fillna(data4['parks'].mean())
data4['residential'] = data4['residential'].fillna(data4['residential'].mean())
data4['retail-and-recreation'] = data4['retail-and-recreation'].fillna(data4['retail-and-recreation'].mean())
data4['transit-stations'] = data4['transit-stations'].fillna(data4['transit-stations'].mean())
data4['workplaces'] = data4['workplaces'].fillna(data4['workplaces'].mean())
data4['isolation'] = data4['isolation'].fillna(data4['isolation'].mean())

data4['federal_district'] = data4['federal_district'].fillna('nan')
data4['geoname_name'] = data4['geoname_name'].fillna('nan')
data4['Start_date'] = data4['Start_date'].dt.dayofyear
data4['urban_pop_rate_c'] = data4['urban_pop_rate_c'].fillna(data4['urban_pop_rate_c'].mean())

russia_regions = russia_regions.rename(columns={'country':'Region'})

regions_ru_vec = pd.read_csv('covid-19-solution/data/features/regions_ru_vec.csv')

data4['Province/State'] = data4['Province/State'].fillna('nan')
regions_ru_vec['Region'] = regions_ru_vec['Region'].fillna('nan')
russia_regions['Province/State'] = russia_regions['Province/State'].fillna('nan')
russia_regions['Region'] = russia_regions['Region'].fillna('nan')

regions_ru_vec = regions_ru_vec.merge(russia_regions[['Region', 'Province/State']], how='left', on='Region')
del regions_ru_vec['Region']

regions_ru_vec_cols = regions_ru_vec.columns[:-1]

data4 = data4.merge(regions_ru_vec, how='left', on='Province/State')

for col in regions_ru_vec_cols:
  data4[col] = data4[col].fillna(data4[col].mean())

#data4.to_csv('covid-19-solution/data/data_with_features/_data_with_features.csv', index=False)

print(data4.shape)
data4.sample(5)

data4.to_csv('covid-19-solution/data/data_with_features/_data_with_features_all.csv', index=False)
