==> 4)_test_features.ipynb <==
# My repo with data
!git clone https://github.com/Tixonmavrin/covid-19-solution
!unzip covid-19-solution/data/data_with_features/_data_with_features.csv.zip
!unzip covid-19-solution/data/data_with_features/_data_with_features_all.csv.zip

import pandas as pd
import numpy as np
import warnings
import copy
import lightgbm as lgb
from sklearn import metrics
from tqdm import tqdm
import random
pd.set_option('display.max_columns', 100)
warnings.filterwarnings("ignore")

#data4 = pd.read_csv('_data_with_features.csv')
data4 = pd.read_csv('_data_with_features_all.csv')

data4['Country/Region'] = data4['Country/Region'].fillna('nan').astype('str')
data4['Province/State'] = data4['Province/State'].fillna('nan').astype('str')
data4['Date'] = pd.to_datetime(data4['Date'])
data4['day'] = data4['Date'].apply(lambda x: x.dayofyear).astype(np.int16)

#data4 = data4[data4['Country/Region'] != 'US']
data4['place'] = data4['Country/Region'].fillna('') + '/' + data4['Province/State'].fillna('')
places = data4['place'].unique()
# Cummax
data4.loc[data4['Country/Region'] != 'Russia','Confirmed'] = data4.loc[data4['Country/Region'] != 'Russia'].groupby("place")["Confirmed"].cummax()
data4.loc[data4['Country/Region'] != 'Russia','Deaths'] = data4.loc[data4['Country/Region'] != 'Russia'].groupby("place")["Deaths"].cummax()

data4n = []
for place in tqdm(data4['place'].unique()):
  if ('Russia' in place) and (',' not in place) and (place.count('Russia') == 1):
    data4n.append(data4[data4['place'] == place])
  else:
    data4c = data4[data4['place'] == place]
    if data4c['Confirmed'].nunique() > 60:
      data4n.append(data4c)
data4 = pd.concat(data4n)

places = data4['place'].unique()
data4['Confirmed per day'] = 0
temp_list = np.zeros(len(data4))
for place in tqdm(places):
    temp = data4['Confirmed'][data4['place']==place].values
    temp[1:] = temp[1:] - temp[:-1]
    data4['Confirmed per day'][data4['place']==place] = temp

def aggregate(data_agg, column, left_right, function, name=''):
    data_agg_new = data_agg.copy()
    column_name = '{}_{}_{}_{}'.format(column, name, left_right[0], left_right[1])
    data_agg_new[column_name] = 0
    tmp = data_agg_new[column].rolling(left_right[1]-left_right[0]+1).agg(function)
    data_agg_new[column_name][left_right[0]:] = tmp[:-left_right[0]]
    return data_agg_new

def make_features(data_agg):
    data_agg = aggregate(data_agg, 'Confirmed per day', [1,1], 'mean', 'mean')
    data_agg = aggregate(data_agg, 'Confirmed per day', [1,7], 'mean','mean')
    data_agg = aggregate(data_agg, 'Confirmed per day', [8,14], 'mean','mean')
    data_agg = aggregate(data_agg, 'Confirmed per day', [15,21], 'mean','mean')

    data_agg = aggregate(data_agg, 'Confirmed per day', [1,7], 'max','max')
    data_agg = aggregate(data_agg, 'Confirmed per day', [1,7], 'min', 'min')
    data_agg = aggregate(data_agg, 'Confirmed per day', [1,7], 'median', 'median')
    
    for thresh in [1, 10, 100]:
        days_under_thresh = (data_agg['Confirmed']<thresh).sum()
        tmp = data_agg['day'].values - days_under_thresh
        tmp[tmp<=0] = 0
        data_agg['days_from_{}'.format(thresh)] = tmp

    for lag in range(1, 14):
        data_agg[f"lag_{lag}_cc"] = data_agg.groupby("place")["Confirmed"].shift(lag)

    data_agg["perc_1_cc"] = data_agg[f"lag_1_cc"] / data_agg.population
    
    data_agg["diff_1_cc"] = data_agg[f"lag_1_cc"] - data_agg[f"lag_2_cc"]
    data_agg["diff_2_cc"] = data_agg[f"lag_2_cc"] - data_agg[f"lag_3_cc"]
    data_agg["diff_3_cc"] = data_agg[f"lag_3_cc"] - data_agg[f"lag_4_cc"]
    
    data_agg["diff_123_cc"] = (data_agg[f"lag_1_cc"] - data_agg[f"lag_4_cc"]) / 3

    data_agg["diff_change_1_cc"] = data_agg.diff_1_cc / data_agg.diff_2_cc
    data_agg["diff_change_2_cc"] = data_agg.diff_2_cc / data_agg.diff_3_cc

    data_agg["diff_change_12_cc"] = (data_agg.diff_change_1_cc + data_agg.diff_change_2_cc) / 2

    data_agg["change_1_cc"] = data_agg[f"lag_1_cc"] / data_agg[f"lag_2_cc"]
    data_agg["change_2_cc"] = data_agg[f"lag_2_cc"] / data_agg[f"lag_3_cc"]
    data_agg["change_3_cc"] = data_agg[f"lag_3_cc"] / data_agg[f"lag_4_cc"]

    data_agg["change_1_3_cc"] = data_agg[f"lag_1_cc"] / data_agg[f"lag_4_cc"]
    data_agg["change_1_7_cc"] = data_agg[f"lag_1_cc"] / data_agg[f"lag_8_cc"]

    data_agg.reset_index(drop=True, inplace=True)
    data_agg["day_from_max"] = 0
    data_agg["max_value"] = 0
    data_agg["delta_with_max"] = 0
    vmax = 0.0
    imax = 0
    for i in range(1, data_agg.shape[0]):
      if data_agg.loc[i-1, 'Confirmed per day'] > vmax:
        vmax = data_agg.loc[i-1, 'Confirmed per day']
        imax = i-1
      data_agg.loc[i, 'delta_with_max'] = data_agg.loc[i,'Confirmed per day'] - vmax
      data_agg.loc[i, 'max_value'] = vmax
      data_agg.loc[i, 'day_from_max'] = i - imax
    
    data_agg["day_from_max_1"] = data_agg.groupby("place")["day_from_max"].shift(1)
    data_agg["max_value_1"] = data_agg.groupby("place")["max_value"].shift(1)
    data_agg["delta_with_max_1"] = data_agg.groupby("place")["delta_with_max"].shift(1)

    data_agg["apl-driving_8"] = data_agg.groupby("place")["apl-driving"].shift(8)
    data_agg["apl-transit_8"] = data_agg.groupby("place")["apl-transit"].shift(8)
    data_agg["apl-walking_8"] = data_agg.groupby("place")["apl-walking"].shift(8)
    data_agg["grocery-and-pharmacy_8"] = data_agg.groupby("place")["grocery-and-pharmacy"].shift(8)
    data_agg["parks_8"] = data_agg.groupby("place")["parks"].shift(8)
    data_agg["residential_8"] = data_agg.groupby("place")["residential"].shift(8)
    data_agg["retail-and-recreation_8"] = data_agg.groupby("place")["retail-and-recreation"].shift(8)
    data_agg["transit-stations_8"] = data_agg.groupby("place")["transit-stations"].shift(8)
    data_agg["workplaces_8"] = data_agg.groupby("place")["workplaces"].shift(8)
    data_agg["isolation_8"] = data_agg.groupby("place")["isolation"].shift(8)

    return data_agg

data5 = []
for place in tqdm(places[:]):
    temp = data4[data4['place']==place].reset_index(drop=True)
    temp = make_features(temp)
    data5.append(temp)
data5 = pd.concat(data5).reset_index(drop=True)

MIN_VALID = '2020-04-20'
MAX_VALID = '2020-04-26'
MIN_TEST = '2020-04-27'
MAX_TEST = '2020-05-03'

data5['Date'] = pd.to_datetime(data5['Date'])
data5.columns = ['Country/Region', 'Province/State', 'Date', 'Confirmed', 'Deaths'] + list(data5.columns[5:])
traint = data5[data5['Date'] < pd.to_datetime(MIN_TEST)]
test = data5[(data5['Date'] >= pd.to_datetime(MIN_TEST)) & (data5['Date'] <= pd.to_datetime(MAX_TEST))]
trainv = data5[data5['Date'] < pd.to_datetime(MIN_VALID)]
valid = data5[(data5['Date'] >= pd.to_datetime(MIN_VALID)) & (data5['Date'] <= pd.to_datetime(MAX_VALID))]
# Add noice
#traint['Confirmed'] = traint['Confirmed'].apply(lambda x: x + random.random())
traint.sample(3)

features = [
 #'Country/Region',
 #'Province/State',
 #'Date',
 #'Confirmed',
 #'Deaths',
 'Health_GDP',
 'Health_USD',
 'Physicians',
 'Nurse',
 'Age_old/new',
 'Smoking',
 'tests',
 'testpop',
 'gatheringlimit',
 'hospibed',
 'healthperpop',
 'density_n',
 'fertility_rate',
 'land_area',
 'median_age',
 'migrants',
 'population_coun',
 'urban_pop_rate_c',
 'world_share',
 'Start_date',
 'population',
 'population_urban',
 'population_rural',
 'urban_pop_rate',
 'lat',
 'lon',
 'continent',
 'area',
 #'apl-driving',
 #'apl-transit',
 #'apl-walking',
 #'grocery-and-pharmacy',
 #'parks',
 #'residential',
 #'retail-and-recreation',
 #'transit-stations',
 #'workplaces',
 #'isolation',
 'federal_district',
 'geoname_name',
 'global_vec_0',
 'global_vec_1',
 'global_vec_2',
 'global_vec_3',
 'global_vec_4',
 'global_vec_5',
 'global_vec_6',
 'global_vec_7',
 'global_vec_8',
 'global_vec_9',
 'global_vec_10',
 'global_vec_11',
 'global_vec_12',
 'global_vec_13',
 'global_vec_14',
 'global_vec_15',
 'global_vec_16',
 'global_vec_17',
 'global_vec_18',
 'global_vec_19',
 'global_vec_20',
 'global_vec_21',
 'global_vec_22',
 'global_vec_23',
 'global_vec_24',
 'global_vec_25',
 'global_vec_26',
 'global_vec_27',
 'region_vec_0',
 'region_vec_1',
 'region_vec_2',
 'region_vec_3',
 'region_vec_4',
 'region_vec_5',
 'region_vec_6',
 'region_vec_7',
 'region_vec_8',
 'region_vec_9',
 'region_vec_10',
 'region_vec_11',
 'region_vec_12',
 'region_vec_13',
 'region_vec_14',
 'region_vec_15',
 'region_vec_16',
 'region_vec_17',
 'region_vec_18',
 'region_vec_19',
 'region_vec_20',
 'region_vec_21',
 'region_vec_22',
 'region_vec_23',
 'region_vec_24',
 'region_vec_25',
 'region_vec_26',
 'region_vec_27',
 'day',
 #'place',
 #'Confirmed per day',
 'Confirmed per day_mean_1_1',
 'Confirmed per day_mean_1_7',
 'Confirmed per day_mean_8_14',
 'Confirmed per day_mean_15_21',
 'Confirmed per day_max_1_7',
 'Confirmed per day_min_1_7',
 'Confirmed per day_median_1_7',
 'days_from_1',
 'days_from_10',
 'days_from_100',
 'lag_1_cc',
 'lag_2_cc',
 'lag_3_cc',
 'lag_4_cc',
 'lag_5_cc',
 'lag_6_cc',
 'lag_7_cc',
 'lag_8_cc',
 'lag_9_cc',
 'lag_10_cc',
 'lag_11_cc',
 'lag_12_cc',
 'lag_13_cc',
 'perc_1_cc',
 'diff_1_cc',
 'diff_2_cc',
 'diff_3_cc',
 'diff_123_cc',
 'diff_change_1_cc',
 'diff_change_2_cc',
 'diff_change_12_cc',
 'change_1_cc',
 'change_2_cc',
 'change_3_cc',
 'change_1_3_cc',
 'change_1_7_cc',
 #'day_from_max',
 #'max_value',
 #'delta_with_max',
 'day_from_max_1',
 'max_value_1',
 'delta_with_max_1',
 'apl-driving_8',
 'apl-transit_8',
 'apl-walking_8',
 'grocery-and-pharmacy_8',
 'parks_8',
 'residential_8',
 'retail-and-recreation_8',
 'transit-stations_8',
 'workplaces_8',
 'isolation_8']

params_sample = {
    'objective': 'regression',
    'metric': 'mse',
    'boosting': 'gbdt',
    'verbosity': 100,
    'bagging_seed': 2020,
    'random_state': 2020,
    'num_round':1000,
    'learning_rate': 0.02,
    'min_data_in_leaf': 5,
    'max_depth': 8,
    'reg_alpha': 1,
    'reg_lambda': 5,}

col_target = 'Confirmed per day'
col_features = [
  'lat', 'lon',
 'days_from_10',
 'Confirmed per day_mean_1_1',
 'Confirmed per day_mean_1_7',
 'Confirmed per day_mean_8_14',
 'Confirmed per day_mean_15_21',
]
col_cat = []

dfc = valid.copy()[['Country/Region','Province/State','Date','Confirmed']]
dfc = dfc[(dfc['Country/Region'] == 'Russia') & (dfc['Province/State'] != 'nan') & (dfc['Province/State'].apply(lambda x: ',' not in x))]

# For add
def mean_absolute_logarithmic_error(df_actual, df_predicted):
    actual = df_actual.astype(np.float).values[:] + 1
    predicted = df_predicted.astype(np.float).values[:] + 1
    return sum(abs(np.log10(predicted) - np.log10(actual)))

def calc_score(pred, df):
  dfc = df.copy()[['Country/Region','Province/State','Date','Confirmed per day']]
  dfc['pred'] = pred
  dfc = dfc[(dfc['Country/Region'] == 'Russia') & (dfc['Province/State'] != 'nan') & (dfc['Province/State'].apply(lambda x: ',' not in x))]
  del dfc['Country/Region']
  dfc = dfc.fillna('nan')
  dfc['Date'] = dfc['Date'].apply(lambda x: str(x))
  all_country = list(dfc['Province/State'].unique())
  dfc.set_index(['Province/State', 'Date'], inplace=True)
  contry_loss = []
  for country in all_country:
      sc = mean_absolute_logarithmic_error(
          dfc.loc[country]['Confirmed per day'],
          dfc.loc[country]['pred']
      )
      contry_loss.append(sc)
  return np.mean(contry_loss)

def check_new_feature(old_features, new_feature, params, train, valid, target_col, cat_col, features_add):
  trainc = train.copy()
  validc = valid.copy()

  X_train = trainc[old_features+[new_feature]+features_add]
  y_train = np.log(trainc[target_col].values.clip(0, 1e10)+1)
  train_data = lgb.Dataset(X_train, label=y_train, categorical_feature=cat_col)

  X_valid = validc[old_features+[new_feature]+features_add]
  y_valid = np.log(validc[target_col].values.clip(0, 1e10)+1)
  valid_data = lgb.Dataset(X_valid, label=y_valid, categorical_feature=cat_col)
  model_pri = lgb.train(params, train_data)

  y_pred = np.exp(model_pri.predict(X_valid))-1
  score1 = calc_score(y_pred, validc)

  trainc[new_feature] = np.random.permutation(trainc[new_feature].values)
  X_train = trainc[old_features+[new_feature]+features_add]
  y_train = np.log(trainc[target_col].values.clip(0, 1e10)+1)
  train_data = lgb.Dataset(X_train, label=y_train, categorical_feature=cat_col)

  X_valid = validc[old_features+[new_feature]+features_add]
  y_valid = np.log(validc[target_col].values.clip(0, 1e10)+1)
  valid_data = lgb.Dataset(X_valid, label=y_valid, categorical_feature=cat_col)
  model_pri = lgb.train(params, train_data)

  y_true = validc[target_col].values
  y_pred = np.exp(model_pri.predict(X_valid))-1
  score2 = calc_score(y_pred, validc)

  print(new_feature, score1, score2, score2 - score1)
  return score2 - score1

col_features2 = col_features
for feature in features:
  try:
    r = check_new_feature(col_features2, feature, params_sample, trainv, valid, col_target, col_cat, [])
    if r < -0.02:
      col_features2.append(feature)
      print('add', feature)
  except Exception as ex:
    print(ex)
    pass
